+++ bpdm-gate/src/main/kotlin/org/eclipse/tractusx/bpdm/gate/service/OutputSaasMappingService.kt
@@ -23,20 +23,94 @@ import mu.KotlinLogging
 import org.eclipse.tractusx.bpdm.common.dto.saas.AugmentedBusinessPartnerResponseSaas
 import org.eclipse.tractusx.bpdm.common.dto.saas.BusinessPartnerSaas
 import org.eclipse.tractusx.bpdm.common.dto.saas.SharingStatusSaas
+import org.eclipse.tractusx.bpdm.common.dto.saas.isError
 import org.eclipse.tractusx.bpdm.common.service.SaasMappings
 import org.eclipse.tractusx.bpdm.gate.config.SaasConfigProperties
 import org.eclipse.tractusx.bpdm.gate.dto.response.ErrorInfo
 import org.eclipse.tractusx.bpdm.gate.exception.BusinessPartnerOutputError
 import org.eclipse.tractusx.bpdm.gate.filterNotNullKeys
+import org.eclipse.tractusx.bpdm.gate.model.BusinessPartnerSaasWithBpn
+import org.eclipse.tractusx.bpdm.gate.model.BusinessPartnerSaasWithExternalId
+import org.eclipse.tractusx.bpdm.gate.model.SharingStatusEvaluationResult
 import org.springframework.stereotype.Service
 import java.time.LocalDateTime
 
 @Service
 class OutputSaasMappingService(
-    private val saasConfigProperties: SaasConfigProperties,
+    private val saasConfigProperties: SaasConfigProperties
 ) {
 
     private val logger = KotlinLogging.logger { }
+
+    fun mapWithLocalBpn(
+        partnersWithExternalId: Collection<BusinessPartnerSaasWithExternalId>,
+        augmentedPartners: Collection<AugmentedBusinessPartnerResponseSaas>
+    ): Collection<BusinessPartnerSaasWithBpn> {
+        val bpnByExternalIdMap = buildBpnByExternalIdMap(augmentedPartners)
+
+        val partnersWithLocalBpn = partnersWithExternalId
+            .filter { partner -> bpnByExternalIdMap[partner.externalId] != null }
+            .map { partner -> BusinessPartnerSaasWithBpn(partner.partner, bpnByExternalIdMap[partner.externalId]!!, partner.externalId) }
+
+        return partnersWithLocalBpn
+    }
+
+    fun evaluateSharingStatus(
+        partners: Collection<BusinessPartnerSaas>,
+        partnersWithLocalBpn: Collection<BusinessPartnerSaasWithBpn>,
+        partnersWithPoolBpn: Collection<BusinessPartnerSaasWithBpn>
+    ): SharingStatusEvaluationResult {
+
+        val partnersWithExternalId = mapWithExternalId(partners)
+
+        val localBpnByExternalId = partnersWithLocalBpn.associateBy { it.externalId }
+        val poolBpnByExternalId = partnersWithPoolBpn.associateBy { it.externalId }
+
+        /// We sort all the entries in one of 3 buckets: valid content, errors or still pending
+        val validExternalIds = mutableListOf<String>()
+        val errors = mutableListOf<ErrorInfo<BusinessPartnerOutputError>>()
+        val pendingExternalIds = mutableListOf<String>()
+
+        partnersWithExternalId.forEach { partnerWithId ->
+            val partner = partnerWithId.partner
+            val sharingStatus = partner.metadata?.sharingStatus
+            val sharingStatusType = sharingStatus?.status
+            val externalId = partnerWithId.externalId
+            val localBpn = localBpnByExternalId[externalId]?.bpn
+            val poolBpn = poolBpnByExternalId[externalId]?.bpn
+
+            if (sharingStatusType == null || sharingStatusType.isError()) {
+                // ERROR: SharingProcessError
+                errors.add(buildErrorInfoSharingProcessError(externalId, sharingStatus))
+            } else if (localBpn != null) {
+                if (poolBpn != null) {
+                    // OKAY: entry found in pool
+                    validExternalIds.add(externalId)
+                } else {
+                    // ERROR: BpnNotInPool
+                    errors.add(buildErrorInfoBpnNotInPool(externalId, localBpn))
+                }
+            } else if (isSharingTimeoutReached(partner)) {
+                // ERROR: SharingTimeout
+                errors.add(buildErrorInfoSharingTimeout(externalId, partner.lastModifiedAt))
+            } else {
+                pendingExternalIds.add(externalId)
+            }
+        }
+
+        return SharingStatusEvaluationResult(validExternalIds, pendingExternalIds, errors)
+    }
+
+
+    fun mapWithExternalId(partners: Collection<BusinessPartnerSaas>): Collection<BusinessPartnerSaasWithExternalId> {
+        val (partnersWithExternalId, partnersWithoutExternalId) = partners.partition { it.externalId != null }
+        //Should only happen when paginating without filtering for externalIds
+        if (partnersWithoutExternalId.isNotEmpty()) {
+            logger.warn { "Encountered ${partnersWithoutExternalId.size} records without an externalId." }
+        }
+
+        return partnersWithExternalId.map { BusinessPartnerSaasWithExternalId(it, it.externalId!!) }
+    }
 
     fun buildBpnByExternalIdMap(augmentedPartnerWrapperCollection: Collection<AugmentedBusinessPartnerResponseSaas>): Map<String, String?> {
         val augmentedPartnerCollection = augmentedPartnerWrapperCollection.mapNotNull { it.augmentedBusinessPartner }